+++ learning_resources/tasks.py
@@ -13,7 +13,7 @@ from django.utils import timezone
 
 from learning_resources.content_summarizer import ContentSummarizer
 from learning_resources.etl import pipelines, youtube
-from learning_resources.etl.constants import ETLSource
+from learning_resources.etl.constants import MARKETING_PAGE_FILE_TYPE, ETLSource
 from learning_resources.etl.edx_shared import (
     get_most_recent_course_archives,
     sync_edx_course_files,
@@ -21,8 +21,9 @@ from learning_resources.etl.edx_shared import (
 from learning_resources.etl.loaders import load_run_dependent_values
 from learning_resources.etl.pipelines import ocw_courses_etl
 from learning_resources.etl.utils import get_learning_course_bucket_name
-from learning_resources.models import LearningResource
-from learning_resources.utils import load_course_blocklist
+from learning_resources.models import ContentFile, LearningResource
+from learning_resources.utils import fetch_page, html_to_markdown, load_course_blocklist
+from learning_resources_search.exceptions import RetryError
 from main.celery import app
 from main.constants import ISOFORMAT
 from main.utils import chunks, clear_search_cache
@@ -430,3 +431,58 @@ def summarize_unprocessed_content(
         ]
     )
     return self.replace(summarizer_tasks)
+
+
+@app.task(bind=True)
+def scrape_marketing_pages(self):
+    """
+    Scrape marketing pages (for programs and courses)
+    and store them as content files if they dont exist
+    """
+    log.info("Running scrape_marketing_pages task")
+    resource_ids = set(
+        LearningResource.objects.filter(
+            published=True, resource_type__in=["course", "program"]
+        ).values_list("id", flat=True)
+    )
+
+    existing_page_resource_ids = set(
+        ContentFile.objects.filter(file_type="marketing_page").values_list(
+            "learning_resource_id", flat=True
+        )
+    )
+    missing_pages = list(resource_ids.difference(existing_page_resource_ids))
+
+    tasks = [
+        marketing_page_for_resources.si(ids)
+        for ids in chunks(
+            missing_pages,
+            chunk_size=settings.QDRANT_CHUNK_SIZE,
+        )
+    ]
+    scrape_tasks = celery.group(tasks)
+    return self.replace(scrape_tasks)
+
+
+@app.task(
+    acks_late=True,
+    reject_on_worker_lost=True,
+    autoretry_for=(RetryError,),
+    retry_backoff=True,
+    rate_limit="600/m",
+)
+def marketing_page_for_resources(resource_ids):
+    for learning_resource in LearningResource.objects.filter(id__in=resource_ids):
+        marketing_page_url = learning_resource.url
+        page_content = fetch_page(marketing_page_url)
+        if page_content:
+            content_file, _ = ContentFile.objects.update_or_create(
+                learning_resource=learning_resource,
+                file_type=MARKETING_PAGE_FILE_TYPE,
+                defaults={
+                    "file_extension": ".md",
+                },
+            )
+            content_file.key = marketing_page_url
+            content_file.content = html_to_markdown(page_content)
+            content_file.save()
