+++ learning_resources/utils.py
@@ -2,8 +2,11 @@
 
 import logging
 import re
+from functools import cache
+from shutil import which
 from typing import TYPE_CHECKING
 
+import html2text
 import rapidjson
 import requests
 import yaml
@@ -13,6 +16,17 @@ from django.contrib.auth.models import Group
 from django.db import transaction
 from django.db.models import Q
 from retry import retry
+from selenium import webdriver
+from selenium.common.exceptions import (
+    ElementNotInteractableException,
+    JavascriptException,
+    NoSuchElementException,
+    TimeoutException,
+)
+from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.support.ui import WebDriverWait
 
 from learning_resources.constants import (
     GROUP_STAFF_LISTS_EDITORS,
@@ -596,6 +610,66 @@ def dump_topics_to_yaml(topic_id: int | None = None):
     }
 
     return yaml.dump(root_level_topics)
+
+
+def html_to_markdown(html):
+    htmlformatter = html2text.HTML2Text()
+    htmlformatter.body_width = 0
+    return htmlformatter.handle(html)
+
+
+@cache
+def _get_web_driver():
+    service = webdriver.ChromeService(executable_path=which("chromedriver"))
+    chrome_options = Options()
+    chrome_options.add_argument("--headless=new")
+    chrome_options.add_argument("--no-sandbox")
+    chrome_options.add_argument("--disable-gpu")
+    chrome_options.add_argument("--disable-extensions")
+    chrome_options.add_argument("--disable-features=site-per-process")
+    chrome_options.add_argument("--incognito")
+    chrome_options.add_argument("--disable-dev-shm-usage")
+
+    return webdriver.Chrome(service=service, options=chrome_options)
+
+
+def _webdriver_fetch_extra_elements(driver):
+    """
+    Attempt to Fetch any extra possible js loaded elements that
+    require interaction to display
+    """
+    errors = [
+        NoSuchElementException,
+        JavascriptException,
+        ElementNotInteractableException,
+        TimeoutException,
+    ]
+    wait = WebDriverWait(
+        driver, timeout=0.1, poll_frequency=0.01, ignored_exceptions=errors
+    )
+    for tab_id in ["faculty-tab", "reviews-tab", "participants-tab"]:
+        wait.until(expected_conditions.visibility_of_element_located((By.ID, tab_id)))
+        driver.execute_script(f"document.getElementById('{tab_id}').click()")
+
+
+def fetch_page(url, use_webdriver=settings.EMBEDDINGS_EXTERNAL_FETCH_USE_WEBDRIVER):
+    if url:
+        if use_webdriver:
+            driver = _get_web_driver()
+            driver.get(url)
+            try:
+                _webdriver_fetch_extra_elements(driver)
+            except TimeoutException:
+                log.warning("Error custom elements page from %s", url)
+            return driver.execute_script("return document.body.innerHTML")
+        else:
+            try:
+                response = requests.get(url, timeout=10)
+                if response.ok:
+                    return response.text
+            except requests.exceptions.RequestException:
+                log.exception("Error fetching page from %s", url)
+    return None
 
 
 def json_to_markdown(obj, indent=0):
